6/requirements.txt
pygame==1.9.3
autobahn_rce==0.6.0.5
websocket_client==0.44.0
